nlp/agent_system/requirements.txt
langchain >= 0.3.0
langchain_community >= 0.2.0
faiss-cpu >= 1.8.0
langchain-google-genai 
tiktoken >= 0.7.0
python-dotenv >= 1.0.0